feat-#101: move microservice-connection to utils module
