+++ src/microservices/auth/services/user/user.service.ts
@@ -5,7 +5,7 @@ import { RestAuthChangeResponse, RestAuthUserId } from '../../rest/models';
 import { GqlAuthChangeResponse, GqlAuthUserId } from '../../graphql/models';
 import { AuthEmailDto, ChangePasswordDto } from '../../rest/dto';
 import { AuthEmailInput, ChangePasswordInput } from '../../graphql/input';
-import { MicroserviceConnectionService } from '../../../../microservices/microservice-connection/microservice-connection.service';
+import { MicroserviceConnectionService } from '../../../../utils/microservice-connection/microservice-connection.service';
 import { GetRefreshUserDto } from '../../rest/dto';
 
 @Injectable()
