+++ src/microservices/booking/rest/booking.controller.spec.ts
@@ -1,7 +1,7 @@
 import { ClientsModule } from '@nestjs/microservices';
 import { Test, TestingModule } from '@nestjs/testing';
-import { MicroserviceConnectionService } from '../../microservice-connection/microservice-connection.service';
-import { connectMicroservice } from '../../microservice-connection/microservice-connection';
+import { MicroserviceConnectionService } from '../../../utils/microservice-connection/microservice-connection.service';
+import { connectMicroservice } from '../../../utils/microservice-connection/microservice-connection';
 import { BookingService } from '../booking.service';
 import { BookingController } from './booking.controller';
