+++ src/microservices/catalog/catalog.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ClientsModule } from '@nestjs/microservices';
 
-import { connectMicroservice } from '../microservice-connection/microservice-connection';
+import { connectMicroservice } from '../../utils/microservice-connection/microservice-connection';
 import { CatalogControllers } from './rest/controllers';
 import { CatalogServices } from './services';
 import { CatalogMutations } from './graphql/mutations';
 import { CatalogQueries } from './graphql/queries';
-import { MicroserviceConnectionService } from '../microservice-connection/microservice-connection.service';
+import { MicroserviceConnectionService } from '../../utils/microservice-connection/microservice-connection.service';
 
 @Module({
   imports: [ClientsModule.registerAsync([connectMicroservice('catalog')])],
