+++ src/microservices/catalog/services/cars/cars.service.spec.ts
@@ -1,7 +1,7 @@
 import { ClientsModule } from '@nestjs/microservices';
 import { Test, TestingModule } from '@nestjs/testing';
-import { MicroserviceConnectionService } from '../../../../microservices/microservice-connection/microservice-connection.service';
-import { connectMicroservice } from '../../../microservice-connection/microservice-connection';
+import { MicroserviceConnectionService } from '../../../../utils/microservice-connection/microservice-connection.service';
+import { connectMicroservice } from '../../../../utils/microservice-connection/microservice-connection';
 import { CarsService } from './cars.service';
 
 describe('CarsService', () => {
