+++ src/utils/microservice-connection/interfaces/microservice-pattern.interface.ts
